splitting CompatAppBuilder to its own module
